Move type modifying type trait to type_traits.hpp
